requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
idna==2.10
numpy==1.20.2
pandas==1.2.3
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
